refactor authentication flow:  remove signin page and added logic to directly open the google signin popup from landing page
